+++ NEWS.md
@@ -79,6 +79,7 @@ Version 0.15b0 - dev
   use `fit_points`, `control_points`, `knots`  and `weights` attributes
 - BUGFIX: `ezdxf.math.has_clockwise_orientation()` returns `True` for 
   counter-clock wise and vice versa
+- BUGFIX: default color for HATCH is 256 (by layer)
   
 Version 0.14.2 - 2020-10-18
 ---------------------------